allow veths to be attached to linuxbridge
commend out this check.
As we're using veth-pairs to connect an ovs-based neutron to a linuxbridge-based tenks, we need to connect a veth interface as type "direct".

however, the sys/class/net/<veth>/master defaults to the other end of the veth pair, failing this check. Adding it to a bridge works fine anyway, so I've just commented this out.